Issue 13: Implemented custom converters and added unit tests

+++ core/src/test/java/name/pehl/piriti/client/gwttest/animal/Bird.java
@@ -35,5 +35,5 @@ public class Bird extends Animal
 
     // ------------------------------------------------------------------- data
 
-    @Json @Xml double wingspan;
+    @Json @Xml int wingspan;
 }
